+++ lib/hackpad/cli/client.rb
@@ -1,5 +1,5 @@
 require 'reverse_markdown'
-require 'colorize'
+require 'paint'
 
 require_relative 'config'
 require_relative 'api'
@@ -14,22 +14,22 @@ module Hackpad
       def initialize(options, output = STDOUT)
         @output = output
         @options = options
-        Store.prepare @options
-        @config = Config.load @options
+        @config = Config.new @options
+        Store.prepare @config
         Api.prepare @config
         if @options[:plain]
-          load File.expand_path('../plain_colors.rb', __FILE__)
+          Paint.mode = 0
         end
       end
 
-      def sites
-        Store.list_sites.each do |site|
-          table site.name, site.url
+      def workspaces
+        @config.workspaces.each do |s|
+          table s.name, s.site
         end
       end
 
       def stats
-        table 'Site', @config['site'].blue
+        table 'Site', Paint[@config.site, :blue]
         table 'Cached Pads', Store.count_pads
         table 'Last Refresh', Store.last_refresh || 'not refreshed yet'
       end
@@ -37,7 +37,7 @@ module Hackpad
       def search(term, start = 0)
         payload = Api.search(term, start)
         payload.each do |a|
-          @output.puts "#{id_or_url a['id']} - #{unescape(a['title']).yellow}"
+          @output.puts "#{id_or_url a['id']} - #{Paint[unescape(a['title']), :yellow]}"
           @output.puts "   #{extract a['snippet']}"
         end
       end
@@ -63,8 +63,8 @@ module Hackpad
       def info(id)
         pad = Pad.new id
         pad.load 'txt'
-        table 'Id', "#{id}".bold
-        table 'Title', "#{pad.title}".yellow
+        table 'Id', Paint[id, :bold]
+        table 'Title', Paint[pad.title, :yellow]
         table 'URI', "#{@config['site']}/#{id}"
         table 'Chars', "#{pad.chars}"
         table 'Lines', "#{pad.lines}"
@@ -87,7 +87,7 @@ module Hackpad
       private
 
       def padline(pad)
-        "#{(@config['site'] + '/') if @options[:urls]}#{pad.id} - #{pad.title}"
+        "#{(@config.site + '/') if @options[:urls]}#{pad.id} - #{pad.title}"
       end
 
       def unescape(s)
@@ -95,7 +95,7 @@ module Hackpad
       end
 
       def extract(s)
-        unescape(s).gsub(/<b class="hit">([^<]*)<\/b>/) { Regexp.last_match[1].cyan.bold }
+        unescape(s).gsub(/<b class="hit">([^<]*)<\/b>/) { Paint[Regexp.last_match[1], :cyan, :bold] }
       end
 
       def table(key, value)
@@ -103,7 +103,7 @@ module Hackpad
       end
 
       def id_or_url(id)
-        "#{(@config['site'] + '/') if @options[:urls]}#{id.bold}"
+        "#{(@config.site + '/') if @options[:urls]}#{Paint[id, :bold]}"
       end
 
     end